requirements.txt
fastapi>=0.110.3
sqlalchemy>=2.0.32
alembic>=1.13.2
pydantic>=2.7.3
pydantic-settings>=2.4.0
asyncpg>=0.29.0
uvicorn>=0.30.6
pydantic[email]>=2.7.3
passlib>=1.7.2
python-jose>=3.3.0
